Re-add reconnect interval
I removed the reconnect interval not since it was calling connect() on the BLE client, which is not the correct way to reconnect. I didn\'t realize then that this would prevent the ESP from reconnecting should it ever disconnect (e.g. when the light gets turned off). Reintroduce the interval but just start scanning (which will establish the connection again)